Fix a race condition in a test
It's possible for the Realm to be notified of both the write it made to
initialize the schema and the write made on the background thread. With an
active read transaction it would discard the second one, but without one we
don't actually know whether or not it's really a new version, and so the
notification block can potentially be called twice.