Work around Xcode 12b1 shipping a broken version of __bit_reference
https://github.com/llvm/llvm-project/commit/f97936fabd263e3b311e3b8e9ffca4920e4fcff0#diff-da408f0e1315929ea96e97824fd3f134
accidentally broke ABI stability for vector<bool>, which was then fixed in
https://github.com/llvm/llvm-project/commit/a829443cc7359ecf0f2de8f82519f511795675ec.
Xcode 12 beta 1 ships the broken one, which means that linking in a library
built using an older version will explode.

Grabbing the fixed header and using that instead when using a prebuilt core
library works around the problem.