Rework how core is downloaded
Xcode requires that any xcframeworks which are part of a target be present at
the start of the build, and so downloading it as part of the build process
doesn't work. Work around this by adding a skeleton xcframework which does not
contain any headers or libraries to the repo, and then populate that skeleton
with the files downloaded.