Fix spurious incorrect thread exceptions due to threadid reuse
We key the Realm cache on the thread ID, but thread IDs are not actually unique
over the lifetime of the process. This means that the cached Realm may have
been created on a different thread with the same ID, which means that it'll be
bound to a different runloop than the current thread's runloop. Check for this,
and simply open a new Realm instead of reusing the cached Realm when this
happens.